fix style errors
